+++ src/pages/MyPage/index.tsx
@@ -1,38 +1,10 @@
 import React from 'react';
 import Topbar from 'components/Topbar';
 import Navbar from 'components/Navbar';
-// import PATH from 'constants/path';
-// import UserInfo from 'components/MyPage/UserInfo';
-import TPlatform from 'types/platform';
-import useApiQuery from 'hooks/useApiQuery';
 import OptionList from 'components/MyPage/OptionList';
-import UserInfoSection from './UserInfo';
+import UserInfoSection from 'components/MyPage/UserInfo';
+import useProfile from './hooks';
 import * as S from './style';
-
-export interface IUserData {
-  nickname: string;
-  email: string;
-  platform: TPlatform;
-  profileImageUrl: '';
-  postedTreesCount: number;
-  savedTreesCount: number;
-  reviewsCount: number;
-}
-
-const useProfile = () => {
-  const { data, isError, isLoading } = useApiQuery<IUserData>('v1/my');
-
-  // TODO: HTTPError로 바꿔주세요.
-  if (isError) {
-    throw new Error('리뷰를 불러오는 데 오류가 발생했습니다.');
-  }
-
-  if (!data && !isLoading) {
-    throw new Error('데이터가 없습니다!');
-  }
-
-  return data;
-};
 
 export const MyPage = () => {
   const userData = useProfile();
